+++ tests/integration/conftest.py
@@ -43,7 +43,7 @@ from aws_tui.vm.services_protocol import Service, ServiceRegistry
 from aws_tui.vm.settings.s3_connections_vm import S3ConnectionsVM
 
 _S3MOCK_IMAGE = (
-    "adobe/s3mock:5.2.0@sha256:7a37f0d796e81a28b970c892dcae532797014616b3312b467af8f0274ebf0c26"
+    "adobe/s3mock:5.2.2@sha256:e7c36014dcf4c7f0f6bec9de888477d1d9b8f55eceb7d9c1186f70d7aadf81ca"
 )
 
 _AWS_CREDENTIAL_ENV_VARS = (